fix(lint): resolve max-lines violations by extracting modules
Extract code from oversized files to bring them under the 300-line limit:
- docker.ts → docker-network.ts (5 network functions)
- open-project.ts → open-project-ssh.ts (SSH connection helpers)
- create-project.ts → docker-identity.ts (identity conflict detection)
- open-project.test.ts → open-project-ssh.test.ts (SSH test cases)

Also includes linter auto-formatting fixes in command-runner.ts,
docker-compose.ts, and auth-sync.ts.
