+++ packages/app/src/lib/core/templates/docker-compose.ts
@@ -1,9 +1,9 @@
 /* jscpd:ignore-start */
 import {
-  resolveComposeProjectName,
   dockerGitSharedCacheVolumeName,
   dockerGitSharedCodexVolumeName,
   resolveComposeNetworkName,
+  resolveComposeProjectName,
   resolveProjectBootstrapVolumeName,
   type TemplateConfig
 } from "../domain.js"
